feat(worker): move persona extract and query-cache cores into kernel
Keep queue wrappers thin and put post-ingest statement processing behind lib/kernel/worker for the strangler path.
